requirements.txt
antlr4-python3-runtime==4.8
backports.functools-lru-cache==1.5
click==6.6
cloudpickle==1.3.0
colorama==0.4.1
cycler==0.10.0
decorator==4.4.1
enum34==1.1.6
findiff==0.8.0
Flask==0.11.1
Flask-Cors==2.1.2
funcsigs==1.0.2
future==0.15.2
geopy==1.11.0
gpsoauth==0.3.0
itsdangerous==0.24
Jinja2==2.8
joblib==0.14.1
kiwisolver==1.1.0
MarkupSafe==0.23
networkx==2.2
numba==0.47.0
numpy==1.16.5
pbr==3.1.0
peewee==2.8.1
Pillow==6.2.2
protobuf==2.6.1
protobuf-to-dict==0.1.0
PyMySQL==0.9.3
pyparsing==2.4.2
python-dateutil==2.8.0
requests==2.10.0
s2sphere==0.2.4
singledispatch==3.4.0.3
six==1.10.0
sklearn==0.0
sqlparse==0.3.0
stevedore==1.23.0
subprocess32==3.5.4
ta==0.5.11
tabulate==0.8.6
typing==3.7.4.1
virtualenv==15.1.0
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.2
Werkzeug==0.11.10
gunicorn==20.0.4
http://download.pytorch.org/whl/cpu/torch-1.5.0%2Bcpu-cp36-cp36m-linux_x86_64.whl
imutils==0.5.3